+++ Package/analyseBot.py
